requirements.txt
pandas
psycopg2-binary
python-dotenv
requests
apache-airflow[postgres]
pyarrow
apache-airflow